Fixed styles and refactor components.
