+++ client/src/components/auth/Login.css
@@ -75,7 +75,7 @@
   line-height: 18rem;
 }
 
-.header-space {
+.header-space-login {
   height: 18rem;
 }
 
@@ -84,7 +84,7 @@
     line-height: 15rem;
   }
 
-  .header-space {
+  .header-space-login {
     height: 15rem;
   }
 }
@@ -94,7 +94,7 @@
     line-height: 12rem;
   }
 
-  .header-space {
+  .header-space-login {
     height: 12rem;
   }
 }
@@ -104,7 +104,7 @@
     line-height: 10rem;
   }
 
-  .header-space {
+  .header-space-login {
     height: 10rem;
   }
 }
@@ -114,7 +114,7 @@
     line-height: 7rem;
   }
 
-  .header-space {
+  .header-space-login {
     height: 7rem;
   }
 }
