+++ client/src/components/sections/Party/PartyMap.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './../MapStyles.css';
-import { MapHeader } from './../MapHeader';
 import ImageFooter from '../../layout/ImageFooter';
+import MapHeader from '../MapHeader';
 
 const PartyMap = () => {
   return (
